Create README.md
This Script will perform an audit of Replicated geodatabase items by Querying the Replica Dataset ID in the GDB Items Table. Then we loop through the returned records and capture the Physical Name and query the GDB Items table for the Physical Name and the data type not equal to the Replica Dataset ID. Then we capture the Item Type, and  Item type name. Then we write the results to the excel. If a GDB has no replicated datasets nothing is written.